+++ printf.h
@@ -51,7 +51,7 @@ t_format	*init_format(void);
 
 void	print_format(t_format *format);
 char	*handle_content(va_list ap, t_format *format);
-char	*handle_integer(t_format *f, int value);
+char	*handle_integer(t_format *f, intmax_t value);
 char	*handle_unsignedint(t_format *f, unsigned int value);
 char	*handle_string(t_format *f, char *str);
 char	*handle_upperhexa(t_format *f, unsigned int value);
